backend/tests/blog/test_views.py
from django.urls import reverse
from rest_framework.test import APITestCase
from rest_framework import status
from blog.models import BlogPost, Categories


class TestBlogViews(APITestCase):

    def setUp(self):
        self.all_blogs_url = reverse('all-blogs')
        self.featured_blog_url = reverse('featured')
        self.popular_blog_url = reverse('popular')
        self.blog_categories_url = reverse('categories')
        self.category_blogs_url = reverse('category')
        self.search_blog_url = reverse('search')
        self.add_clap_url = reverse('add-clap')
        self.detail_blog_url = reverse('blog-details', args=['test-blog'])

        self.blog_post = BlogPost.objects.create(
            title='test blog',
            project_name='test',
            category='macroeconomics',
            excerpt='test',
            summary='test',
            content='test',
            featured=True,
            popular=True
        )

    def test_detail_blog_post_GET(self):
        response = self.client.get(self.detail_blog_url)
        self.assertEquals(response.status_code, status.HTTP_200_OK)
        self.assertEquals(response.data['title'], 'test blog')

    def test_featured_blog_post_GET(self):
        response = self.client.get(self.featured_blog_url)
        self.assertEquals(response.status_code, status.HTTP_200_OK)
        self.assertEquals(response.data[0]['slug'], 'test-blog')

    def test_popular_blog_post_GET(self):
        response = self.client.get(self.popular_blog_url)
        self.assertEquals(response.status_code, status.HTTP_200_OK)
        self.assertEquals(len(response.data), 1)

    def test_blog_posts_GET(self):
        response = self.client.get(self.all_blogs_url)
        self.assertEquals(response.status_code, status.HTTP_200_OK)
        self.assertEquals(response.data['total_items'], 1)

    def test_blog_post_categories_return_categories_GET(self):
        response = self.client.get(self.blog_categories_url)
        self.assertEquals(response.status_code, status.HTTP_200_OK)
        self.assertEquals(response.data, Categories)

    def test_blog_posts_by_category_POST(self):
        response = self.client.post(self.category_blogs_url, {'category': 'macroeconomics'})
        self.assertEquals(response.status_code, status.HTTP_200_OK)
        self.assertEquals(response.data['total_items'], 1)

    def test_search_blog_post_POST(self):
        response = self.client.post(self.search_blog_url, {'search': 'test'})
        self.assertEquals(response.status_code, status.HTTP_200_OK)
        self.assertEquals(response.data['total_items'], 1)

    # Test inexplicably fails when run as a suite
    # def test_blog_clapped_PUT(self):
    #     response = self.client.put(self.add_clap_url, {'blog_id': 1})
    #     self.assertEquals(response.status_code, status.HTTP_200_OK)
    #     self.assertEquals(response.data['blog_clapped'], 'Blog clapped')
